+++ docs/CHANGE_LOG.md
@@ -6,7 +6,8 @@
 - Added a new binding option called "onRefresh", which states an event that should be triggered when a heatmap is refreshed.
 
 #### **Public Functions:**
-- Added new public function "removeDate()", which will remove a specific date from the heat map (or decrease the date if more than one exists).
+- Added new public function "removeDate()", which will remove a specific date from a heat map (or decrease the date if more than one exists).
+- Added new public function "reset()", which will remove all the dates from a the heat map.
 
 <br>